vault: use new balance allocation

+++ contracts/pools/stable/StablePool.sol
@@ -314,7 +314,7 @@ contract StablePool is IPool, IGeneralPoolQuote, StableMath, BalancerPoolToken, 
     }
 
     function _addSwapFee(uint256 amount) private view returns (uint256) {
-        return amount.div(uint256(FixedPoint.ONE).sub(_swapFee));
+        return amount.div(FixedPoint.ONE.sub(_swapFee));
     }
 
     function _subtractSwapFee(uint256 amount) private view returns (uint256) {
